moved app into components folder, changed imports, and added Navbar
