+++ src/index.js
@@ -1,8 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter as Router } from "react-router-dom";
-import App from "./App";
-import "./app.css";
+import App from "./components/App";
 
 
 ReactDOM.render((<React.StrictMode>
